+++ protoc_plugin/test/goldens/int64.pb.dart
@@ -18,16 +18,16 @@ import 'package:protobuf/protobuf.dart' as $pb;
 export 'package:protobuf/protobuf.dart' show GeneratedMessageGenericExtensions;
 
 class Int64 extends $pb.GeneratedMessage {
-  factory Int64() => create();
+  factory Int64() => Int64._();
 
   Int64._();
 
   factory Int64.fromBuffer($core.List<$core.int> data,
           [$pb.ExtensionRegistry registry = $pb.ExtensionRegistry.EMPTY]) =>
-      create()..mergeFromBuffer(data, registry);
+      Int64._()..mergeFromBuffer(data, registry);
   factory Int64.fromJson($core.String json,
           [$pb.ExtensionRegistry registry = $pb.ExtensionRegistry.EMPTY]) =>
-      create()..mergeFromJson(json, registry);
+      Int64._()..mergeFromJson(json, registry);
 
   static final $pb.BuilderInfo _i = $pb.BuilderInfo(
       _omitMessageNames ? '' : 'Int64',
@@ -45,9 +45,9 @@ class Int64 extends $pb.GeneratedMessage {
   $pb.BuilderInfo get info_ => _i;
 
   @$core.pragma('dart2js:noInline')
-  static Int64 create() => Int64._();
+  static $pb.GeneratedMessage create() => Int64._();
   @$core.override
-  Int64 createEmptyInstance() => create();
+  Int64 createEmptyInstance() => Int64._();
   @$core.pragma('dart2js:noInline')
   static Int64 getDefault() =>
       _defaultInstance ??= $pb.GeneratedMessage.$_defaultFor<Int64>(create);